# packages/contracts/schema/companion.openapi.json
{
  "info": {
    "title": "Voice Director",
    "description": "Local controls for the Slay the Spire 2 voice mod.",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "localToken": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  },
  "security": [
    {
      "localToken": []
    }
  ],
  "openapi": "3.1.2",
  "paths": {
...
    "/api/jobs": {
      "get": {
        "operationId": "listJobs",
        "tags": [
          "Generation"
        ],
        "parameters": [
          {
            "name": "runId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response for status 200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "id",
                      "runId",
                      "kind",
                      "status",
                      "instanceId",
                      "definitionId",
                      "lucky",
                      "immediate",
                      "createdAt",
                      "completedAt",
                      "error",
                      "result"
                    ],
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "runId": {
                        "type": "string"
                      },
                      "kind": {
                        "type": "string",
                        "enum": [
                          "card",
                          "art"
                        ]
                      },
                      "status": {
                        "type": "string"
                      },
                      "instanceId": {
                        "type": "string"
                      },
                      "definitionId": {
                        "nullable": true,
                        "type": [
                          "string",
                          "null"
                        ]
                      },
                      "lucky": {
                        "type": "boolean"
                      },
                      "immediate": {
                        "type": "boolean"
                      },
                      "createdAt": {
                        "type": "string"
                      },
                      "completedAt": {
                        "nullable": true,
                        "type": [
                          "string",
                          "null"
                        ]
                      },
                      "error": {
                        "nullable": true,
                        "type": [
                          "string",
                          "null"
                        ]
                      },
                      "result": {
                        "nullable": true,
                        "type": [
                          "string",
                          "null"
                        ]
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createCardJob",
        "tags": [
          "Generation"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "instanceId",
                  "lucky",
                  "immediate"
                ],
                "properties": {
                  "instanceId": {
                    "type": "string"
                  },
                  "lucky": {
                    "default": false,
                    "type": "boolean"
                  },
                  "immediate": {
                    "default": false,
                    "type": "boolean"
                  }
                }
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "required": [
                  "instanceId",
                  "lucky",
                  "immediate"
                ],
                "properties": {
                  "instanceId": {
                    "type": "string"
                  },
                  "lucky": {
                    "default": false,
                    "type": "boolean"
                  },
                  "immediate": {
                    "default": false,
                    "type": "boolean"
                  }
                }
              }
            },
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "instanceId",
                  "lucky",
                  "immediate"
                ],
                "properties": {
                  "instanceId": {
                    "type": "string"
                  },
                  "lucky": {
                    "default": false,
                    "type": "boolean"
                  },
                  "immediate": {
                    "default": false,
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Response for status 200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "runId",
                    "kind",
                    "status",
                    "instanceId",
                    "definitionId",
                    "lucky",
                    "immediate",
                    "createdAt",
                    "completedAt",
                    "error",
                    "result"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "runId": {
                      "type": "string"
                    },
                    "kind": {
                      "type": "string",
                      "enum": [
                        "card",
                        "art"
                      ]
                    },
                    "status": {
                      "type": "string"
                    },
                    "instanceId": {
                      "type": "string"
                    },
                    "definitionId": {
                      "nullable": true,
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "lucky": {
                      "type": "boolean"
                    },
                    "immediate": {
                      "type": "boolean"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "completedAt": {
                      "nullable": true,
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "error": {
                      "nullable": true,
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "result": {
                      "nullable": true,
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/decisions": {
      "get": {
        "operationId": "listAmbientDecisions",
        "tags": [
          "Voice"
        ],
        "parameters": [
          {
            "name": "runId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response for status 200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "id",
                      "runId",
                      "at",
                      "action",
                      "reason",
                      "anchorId",
                      "transcriptIds",
                      "contextIds",
                      "windowStartedAt",
                      "windowEndedAt",
                      "classifier",
                      "confirmation"
                    ],
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "runId": {
                        "type": "string"
                      },
                      "at": {
                        "type": "string"
                      },
                      "action": {
                        "type": "string",
                        "enum": [
                          "ignore",
                          "create_card"
                        ]
                      },
                      "reason": {
                        "type": "string"
                      },
                      "anchorId": {
                        "type": "string"
                      },
                      "transcriptIds": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "contextIds": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "windowStartedAt": {
                        "type": "string"
                      },
                      "windowEndedAt": {
                        "type": "string"
                      },
                      "classifier": {
                        "additionalProperties": false,
                        "type": [
                          "object",
                          "null"
                        ],
                        "required": [
                          "answers",
                          "inputs"
                        ],
                        "properties": {
                          "answers": {
                            "additionalProperties": false,
                            "type": "object",
                            "required": [
                              "create_card"
                            ],
                            "properties": {
                              "create_card": {
                                "additionalProperties": false,
                                "type": "object",
                                "required": [
                                  "type",
                                  "noul"
                                ],
                                "properties": {
                                  "type": {
                                    "const": "noul",
                                    "type": "string"
                                  },
                                  "noul": {
                                    "minimum": 0,
                                    "maximum": 1,
                                    "type": "number"
                                  }
                                }
                              }
                            }
                          },
                          "inputs": {
                            "minItems": 1,
                            "type": "array",
                            "items": {
                              "additionalProperties": false,
                              "type": "object",
                              "required": [
                                "id",
                                "answers"
                              ],
                              "properties": {
                                "id": {
                                  "type": "string"
                                },
                                "answers": {
                                  "additionalProperties": false,
                                  "type": "object",
                                  "required": [
                                    "noise",
                                    "accidental"
                                  ],
                                  "properties": {
                                    "noise": {
                                      "additionalProperties": false,
                                      "type": "object",
                                      "required": [
                                        "type",
                                        "noul"
                                      ],
                                      "properties": {
                                        "type": {
                                          "const": "noul",
                                          "type": "string"
                                        },
                                        "noul": {
                                          "minimum": 0,
                                          "maximum": 1,
                                          "type": "number"
                                        }
                                      }
                                    },
                                    "accidental": {
                                      "additionalProperties": false,
                                      "type": "object",
                                      "required": [
                                        "type",
                                        "noul"
                                      ],
                                      "properties": {
                                        "type": {
                                          "const": "noul",
                                          "type": "string"
                                        },
                                        "noul": {
                                          "minimum": 0,
                                          "maximum": 1,
                                          "type": "number"
                                        }
                                      }
                                    }
                                  }
                                }
                              }
                            }
                          }
                        }
                      },
                      "confirmation": {
                        "additionalProperties": false,
                        "type": [
                          "object",
                          "null"
                        ],
                        "required": [
                          "answers",
                          "instanceId"
                        ],
                        "properties": {
                          "answers": {
                            "additionalProperties": false,
                            "type": "object",
                            "required": [
                              "create_card"
                            ],
                            "properties": {
                              "create_card": {
                                "additionalProperties": false,
                                "type": "object",
                                "required": [
                                  "type",
                                  "noul"
                                ],
                                "properties": {
                                  "type": {
                                    "const": "noul",
                                    "type": "string"
                                  },
                                  "noul": {
                                    "minimum": 0,
                                    "maximum": 1,
                                    "type": "number"
                                  }
                                }
                              }
                            }
                          },
                          "instanceId": {
                            "type": [
                              "string",
                              "null"
                            ]
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{id}/cancel": {
      "post": {
        "operationId": "cancelJob",
        "tags": [
          "Generation"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response for status 200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "runId",
                    "kind",
                    "status",
                    "instanceId",
                    "definitionId",
                    "lucky",
                    "immediate",
                    "createdAt",
                    "completedAt",
                    "error",
                    "result"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "runId": {
                      "type": "string"
                    },
                    "kind": {
                      "type": "string",
                      "enum": [
                        "card",
                        "art"
                      ]
                    },
                    "status": {
                      "type": "string"
                    },
                    "instanceId": {
                      "type": "string"
                    },
                    "definitionId": {
                      "nullable": true,
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "lucky": {
                      "type": "boolean"
                    },
                    "immediate": {
                      "type": "boolean"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "completedAt": {
                      "nullable": true,
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "error": {
                      "nullable": true,
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "result": {
                      "nullable": true,
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/transcripts": {
      "get": {
        "operationId": "listTranscripts",
        "tags": [
          "Voice"
        ],
        "parameters": [
          {
            "name": "runId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response for status 200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "id",
                      "runId",
                      "text",
                      "source",
                      "mood",
                      "createdAt"
                    ],
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "runId": {
                        "type": "string"
                      },
                      "text": {
                        "type": "string"
                      },
                      "source": {
                        "type": "string"
                      },
                      "mood": {
                        "type": "string"
                      },
                      "createdAt": {
                        "type": "string"
                      },
                      "playerId": {
                        "type": "string"
                      },
                      "playerName": {
                        "type": "string"
                      },
                      "startedAt": {
                        "type": "string"
                      },
                      "endedAt": {
                        "type": "string"
                      },
                      "noiseProbability": {
                        "type": "number"
                      },
                      "accidentalProbability": {
                        "type": "number"
                      },
                      "error": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/art/{id}": {
      "get": {
        "operationId": "getArtwork",
        "tags": [
          "Cards"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  }
}